Continued plaform compatibility fixes (64-bit vs 32-bit MSVC)

+++ SmileC/smilelibtests/ansicolor-w32.c
@@ -296,7 +296,7 @@ retry:
       ptr++;
     }
   }
-  return ptr - buf;
+  return (int)(ptr - buf);
 }
 
 int
